fix smoother dimension
@cekees:
I'm almost positive this is correct...
I kept getting errors from numpy's
concatenate function when making
small changes to the configuration of this
test case.

My investigation shows that the current
setup generates a "lucky" number of nodes
such that the velocityDOF arrays in
the SimpleNavierStokes3D preconditioner are
the right size.
However, when this does work out the
case is probably getting very bad
preconditioning and when it doesn't
work out it just won't run.