First tests of dambreak_Colagrossi
Code can run, but the parameters, boundary conditions, and the like need to be pinned down more (most are based off of the Coarse test case, but not all have direct analogues and some conditions are not yet checked to match exactly - and the coarse test case never ran for me in master due to other issues, so it could be a problem in its own right).  Gauges are currently not included as their errors are hard to determine the blame for (whether it is an issue with the gauge logic or one of the cases where Proteus' gauges break), but the code is there to test.