Added in boundary conditions.
Now that the rest of the code is stable, these boundary conditions allow the test to be run with actual results (and not just the no boundary conditions set warning and solver failure).  I'm editing this from github, but I originally made these changes locally just before I left yesterday.