First setup for refactor.
Old files are temporarily in the new directory constantSlopeBeach_old, which is deprecated.  The new file structure will be rebuilt in the new replacement constantSlopeBreach directory (which will be fleshed out in the next few commits).  This also brought in the waves of Xie (a vertical breakwater problem) and its readme file.