Directory/file setup for all constant slope beach problems
Including adding a copy of the old Waves_Xie code to constantSlopeBeach_old for reference usage.  Mase and Kirby shoaling was already set up, and mostly adapted, although there is a major bug there as of this commit.  The others just have their old code at the moment - this update was to get the right files in the right places for adaption.