Deprecated variant tests
Sharp crested weir variants moved to a special subdirectory to emphasize that they are deprecated.  They remain with the code so that context files can be made off of them, since sharp crested weir has a lot of different (and not trivially different) variants which will need to be carefully looked at while making context files.  Once context files cover one of these variant test concepts, it should be removed until the deprecated folder is empty.