Cleanup and variants folders
As with the others, the default.txt files are just a placeholder.  Test context option files should go there (with probably some form of default test - it just has to have some values in it or else parun will reject it), while all variant context option choices can be put in the main variants folder in each test.